+++ packages/core/src/index.ts
@@ -33,6 +33,8 @@ export {
 } from './form';
 export { RequestError, initErrorHandle } from './global/error';
 export { downloadInterceptor } from './global/interceptors';
+export { InfiniteList } from './infinite_list';
+export type { InfiniteListProps } from './infinite_list';
 export { Markdown } from './markdown';
 export { PageLayout } from './page_layout';
 export { routeTool } from './route';