+++ packages/tool/src/index.ts
@@ -2,5 +2,6 @@ export { copyToClipboard } from './copy';
 export { getDayjs } from './date';
 export { Environment } from './environment';
 export { pinyin, pinyinFilter, pinyinMatch, pinyinMatchWithoutFirstLetter } from './pinyin';
+export { sleep } from './sleep';
 export { buildURL } from './url';
 export { useGlobalRequest } from './use_global_request';